Backend/api/requirements.txt
fastapi==0.111.0
uvicorn==0.21.1
python-multipart==0.0.9
python-dotenv==1.0.0
pydantic==2.7.4
httpx==0.24.1 
itsdangerous==2.2.0
urllib3==2.4.0
starlette==0.37.2